Interface and good tests
The music_bubbles package now has an interface that depends on
underlying modules. Those modules _can_ be accessed by outside code, but
only if you peek at the implementation.
The only public function takes a list of note names as strings and
returns a list of dictionaries that contain the name, diameter, xy
positions, and color of each note.